requirements.txt
fastapi
uvicorn
pinecone==3.1.0
tiktoken
python-dotenv
pydantic
langchain-openai
